future 异步方法

+++ commons/gpromise/promise.go
@@ -5,7 +5,6 @@ import (
 	"container/list"
 	"errors"
 	"fmt"
-	"log"
 	"runtime"
 	"strconv"
 	"strings"
@@ -70,6 +69,7 @@ type Promise struct {
 	cbFunc        func(context *Context)
 	context       *Context
 	futureIdIndex uint32
+	logger        promiseLogger
 }
 
 type Callback struct {
@@ -89,9 +89,11 @@ type Manager struct {
 	warnThresholdFunc    func() int // 获取需要打印警告的promise耗时时间
 
 	onAllPromiseOver func()
+
+	logger promiseLogger
 }
 
-func NewManager(owner int64, f func() int) *Manager {
+func NewManager(owner int64, f func() int, logger promiseLogger) *Manager {
 	pm := &Manager{
 		owner:                owner,
 		promises:             make(map[uint32]*Promise),
@@ -99,6 +101,7 @@ func NewManager(owner int64, f func() int) *Manager {
 		statResults:          []*StatResult{},
 		lastPrintPromiseStat: time.Now(),
 		warnThresholdFunc:    f,
+		logger:               logger,
 	}
 
 	return pm
@@ -117,6 +120,7 @@ func (pm *Manager) NewPromise(name string, cbFunc func(context *Context), args .
 			trace: strings.Builder{},
 		},
 		futureIdIndex: 1,
+		logger:        pm.logger,
 	}
 	p.context.Trace(name, 0)
 	pm.promiseIdIndex++
@@ -164,12 +168,12 @@ func (pm *Manager) DeletePromise(Id uint32) {
 
 func (p *Promise) Push(future IFuture) {
 	if future == nil {
-		log.Println("Error: future is nil")
+		p.logger.PromiseErrorF("Promise: future is nil")
 		return
 	}
 
 	if p == nil {
-		log.Println("Error: promise is nil")
+		p.logger.PromiseErrorF("Promise: promise is nil")
 		return
 	}
 
@@ -221,8 +225,8 @@ func (p *Promise) Start() {
 		if er != nil {
 			if !errors.Is(er, ContinueErr) {
 				p.context.Err = er
-				log.Println(fmt.Sprintf("Warn: Owner[%v] promise[%v] future[%v] %v CallBack fail[%v]",
-					p.pm.owner, p.Id, f.Id(), GetPfId(p.Id, f.Id()), p.context.Err.Error()))
+				p.logger.PromiseWarnF("Promise: Owner[%v] promise[%v] future[%v] %v do fail[%v]",
+					p.pm.owner, p.Id, f.Id(), GetPfId(p.Id, f.Id()), p.context.Err.Error())
 				promiseFinish = true
 				break
 			} else {
@@ -243,10 +247,13 @@ func (p *Promise) Start() {
 	}
 
 	if promiseFinish {
-		if p.cbFunc != nil {
-			p.cbFunc(p.context)
+		// 检查是否仍然存在于 Manager 中
+		if _, exists := p.pm.promises[p.Id]; exists {
+			if p.cbFunc != nil {
+				p.cbFunc(p.context)
+			}
+			p.pm.DeletePromise(p.Id)
 		}
-		p.pm.DeletePromise(p.Id)
 	}
 }
 
@@ -263,22 +270,22 @@ func (pm *Manager) Process(pfId uint64, args []interface{}, errInfo error) {
 
 	p, ok := pm.promises[promiseId]
 	if !ok {
-		log.Println(fmt.Sprintf("Warn: owner[%v] promise[%v] future[%v] %v is not exist",
-			pm.owner, promiseId, futureId, pfId))
+		p.logger.PromiseWarnF("Manager: Owner[%v] promise[%v] future[%v] %v is not exist",
+			pm.owner, promiseId, futureId, pfId)
 		return
 	}
 
 	e, ok := p.futureMap[futureId]
 	if !ok {
-		log.Println(fmt.Sprintf("Error: owner[%v] promise[%v] future[%v] %v is not exist",
-			pm.owner, promiseId, futureId, pfId))
+		p.logger.PromiseErrorF("Manager: Owner[%v] promise[%v] future[%v] %v is not exist",
+			pm.owner, promiseId, futureId, pfId)
 		return
 	}
 
 	f, ok := e.Value.(IFuture)
 	if !ok {
-		log.Println(fmt.Sprintf("Error: owner[%v] promise[%v] future[%v] %v element is not future",
-			pm.owner, promiseId, futureId, pfId))
+		p.logger.PromiseErrorF("Manager: Owner[%v] promise[%v] future[%v] %v element is not future",
+			pm.owner, promiseId, futureId, pfId)
 		p.context.Err = errors.New("element is not future")
 		if p.cbFunc != nil {
 			p.cbFunc(p.context)
@@ -288,8 +295,8 @@ func (pm *Manager) Process(pfId uint64, args []interface{}, errInfo error) {
 	}
 
 	if errInfo != nil {
-		log.Println(fmt.Sprintf("Warn: owner[%v] promise[%v] future[%v] %v do fail[%v]",
-			pm.owner, promiseId, futureId, pfId, errInfo.Error()))
+		p.logger.PromiseWarnF("Manager: Owner[%v] promise[%v] future[%v] %v do fail[%v]",
+			pm.owner, promiseId, futureId, pfId, errInfo.Error())
 		p.context.Err = errInfo
 		if p.cbFunc != nil {
 			p.cbFunc(p.context)
@@ -310,8 +317,8 @@ func (pm *Manager) Process(pfId uint64, args []interface{}, errInfo error) {
 	if er != nil {
 		if !errors.Is(er, ContinueErr) {
 			p.context.Err = er
-			log.Println(fmt.Sprintf("Warn: owner[%v] promise[%v] future[%v] %v CallBack fail[%v]",
-				pm.owner, promiseId, futureId, pfId, p.context.Err.Error()))
+			p.logger.PromiseWarnF("Manager: Owner[%v] promise[%v] future[%v] %v CallBack fail[%v]",
+				pm.owner, promiseId, futureId, pfId, p.context.Err.Error())
 			if p.cbFunc != nil {
 				p.cbFunc(p.context)
 			}
@@ -352,8 +359,8 @@ func (pm *Manager) Process(pfId uint64, args []interface{}, errInfo error) {
 		if er != nil {
 			if !errors.Is(er, ContinueErr) {
 				p.context.Err = er
-				log.Println(fmt.Sprintf("Warn: owner[%v] promise[%v] future[%v] %v CallBack fail[%v]",
-					pm.owner, p.Id, f.Id(), pfId, p.context.Err.Error()))
+				p.logger.PromiseWarnF("Manager: Owner[%v] promise[%v] future[%v] %v do fail[%v]",
+					pm.owner, p.Id, f.Id(), pfId, p.context.Err.Error())
 				promiseFinish = true
 				break
 			} else {
@@ -390,7 +397,7 @@ func (pm *Manager) Update(during time.Duration) {
 				p.cbFunc(p.context)
 			}
 			p.pm.DeletePromise(p.Id)
-			log.Println(fmt.Sprintf("Error: owner[%v] promise[%v][%v][%v] run timeout, delete", pm.owner, p.Id, p.Name, p.GetAllPfId()))
+			pm.logger.PromiseErrorF("Manager: Owner[%v] promise[%v][%v] run timeout, delete", pm.owner, p.Id, p.Name)
 		}
 	}
 
@@ -433,13 +440,13 @@ func (pm *Manager) printStat() {
 	count := len(pm.statResults)
 
 	if len(pm.statResults) > 0 {
-		log.Println(fmt.Sprintf("Info: Owner[%v] PromiseStatus [PromiseCount(%v), AvgPromiseDuration(%v), MaxPromiseDuration(%v) PromiseName(%v)",
-			pm.owner, count, getDurationDesc(avgDuration), getDurationDesc(maxDuration), pm.maxResult.Name))
+		pm.logger.PromiseInfoF("Manager: Owner[%v] PromiseStatus [PromiseCount(%v), AvgPromiseDuration(%v), MaxPromiseDuration(%v) PromiseName(%v)",
+			pm.owner, count, getDurationDesc(avgDuration), getDurationDesc(maxDuration), pm.maxResult.Name)
 		if pm.warnThresholdFunc != nil {
 			thresholdValue := pm.warnThresholdFunc()
 			if thresholdValue != 0 && maxDuration >= time.Duration(thresholdValue)*time.Millisecond {
-				log.Println(fmt.Sprintf("Warn: Owner[%v] PromiseName %v timeout excute %v exceed limit %v",
-					pm.owner, pm.maxResult.Name, getDurationDesc(pm.maxResult.Duration), thresholdValue))
+				pm.logger.PromiseWarnF("Manager: Owner[%v] PromiseName %v timeout excute %v exceed limit %v",
+					pm.owner, pm.maxResult.Name, getDurationDesc(pm.maxResult.Duration), thresholdValue)
 			}
 		}
 	}
